refactor: move calculating races from activity; fix: junior age while generate rower
